+++ sources/packages/web/src/views/student/StudentApplicationDetails.vue
@@ -75,19 +75,7 @@
         :applicationDetails="applicationDetails"
       />
     </v-container>
-    <request-assessment
-      :applicationId="id"
-      :showWhenEmpty="false"
-      @viewStudentAppeal="goToStudentAppeal"
-      @viewApplicationException="goToApplicationException"
-      @viewOfferingRequest="goToOfferingRequest"
-    />
-    <history-assessment
-      :applicationId="id"
-      :viewRequestTypes="studentAssessmentRequestTypes"
-      @viewStudentAppeal="goToStudentAppeal"
-      @viewAssessment="gotToViewAssessment"
-    />
+    <student-assessment-details :applicationId="id" v-if="showViewAssessment" />
   </student-page-container>
   <ConfirmEditApplication ref="editApplicationModal" />
 </template>
@@ -99,24 +87,17 @@ import CancelApplication from "@/components/students/modals/CancelApplicationMod
 import { ApplicationService } from "@/services/ApplicationService";
 import "@/assets/css/student.scss";
 import { useFormatters, ModalDialog, useSnackBar } from "@/composables";
-import {
-  GetApplicationDataDto,
-  ApplicationStatus,
-  MenuType,
-  AssessmentTriggerType,
-} from "@/types";
+import { GetApplicationDataDto, ApplicationStatus, MenuType } from "@/types";
 import ApplicationDetails from "@/components/students/ApplicationDetails.vue";
 import ConfirmEditApplication from "@/components/students/modals/ConfirmEditApplication.vue";
-import HistoryAssessment from "@/components/aest/students/assessment/History.vue";
-import RequestAssessment from "@/components/aest/students/assessment/Request.vue";
+import StudentAssessmentDetails from "@/components/students/StudentAssessmentDetails.vue";
 
 export default {
   components: {
     CancelApplication,
     ApplicationDetails,
     ConfirmEditApplication,
-    HistoryAssessment,
-    RequestAssessment,
+    StudentAssessmentDetails,
   },
   props: {
     id: {
@@ -132,10 +113,6 @@ export default {
     const applicationDetails = ref({} as GetApplicationDataDto);
     const editApplicationModal = ref({} as ModalDialog<boolean>);
     const snackBar = useSnackBar();
-    // The assessment trigger types for which the request form is visible to student.
-    const studentAssessmentRequestTypes = ref<AssessmentTriggerType[]>([
-      AssessmentTriggerType.StudentAppeal,
-    ]);
 
     const showHideCancelApplication = () => {
       showModal.value = !showModal.value;
@@ -238,25 +215,6 @@ export default {
         await ApplicationService.shared.getApplicationData(applicationId);
       loadMenu();
     };
-    const gotToViewAssessment = (assessmentId: number) => {
-      router.push({
-        name: StudentRoutesConst.ASSESSMENT_AWARD_VIEW,
-        params: {
-          applicationId: props.id,
-          assessmentId,
-        },
-      });
-    };
-
-    const goToStudentAppeal = (appealId: number) => {
-      router.push({
-        name: StudentRoutesConst.STUDENT_APPEAL_REQUESTS,
-        params: {
-          applicationId: props.id,
-          appealId,
-        },
-      });
-    };
 
     watch(
       () => props.id,
@@ -283,9 +241,6 @@ export default {
       editApplicationModal,
       editApplication,
       viewApplication,
-      studentAssessmentRequestTypes,
-      gotToViewAssessment,
-      goToStudentAppeal,
     };
   },
 };
